+++ Dockerfile
@@ -3,7 +3,7 @@ FROM tomcat:10-jdk17-corretto
 ENV TZ=Asia/Seoul
 
 RUN rm -rf /usr/local/tomcat/webapps/*
-COPY hiapp/build/libs/hifive.war /usr/local/tomcat/webapps/ROOT.war
+COPY bururung/build/libs/hifive.war /usr/local/tomcat/webapps/ROOT.war
 
 EXPOSE 8080
