+++ hiapp/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'hiapp'
+rootProject.name = 'bururung'
