+++ hiapp/src/main/java/com/hifive/bururung/BururungApplication.java
@@ -1,0 +1,14 @@
+package com.hifive.bururung;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class BururungApplication extends SpringBootServletInitializer {
+	
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
+		return builder.sources(BururungApplication.class);
+	}
+}
