+++ hiapp/src/main/java/com/hifive/bururung/domain/member/dto/LoginRequest.java
@@ -1,0 +1,17 @@
+package com.hifive.bururung.domain.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginRequest {
+
+	private String email;
+	
+	private String password;
+}
