+++ hiapp/src/main/java/com/hifive/bururung/domain/member/dto/SignupRequest.java
@@ -1,0 +1,28 @@
+package com.hifive.bururung.domain.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignupRequest {
+	
+	private String email;
+	
+	private String password;
+	
+	private String phone;
+	
+	private String name;
+	
+	private Character gender;
+	
+	private String birth;
+	
+	private String nickname;
+}
