+++ hiapp/src/main/java/com/hifive/bururung/domain/member/dto/TokenDTO.java
@@ -1,0 +1,19 @@
+package com.hifive.bururung.domain.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenDTO {
+	
+	private String accessToken;
+	
+	private String refreshToken;
+	
+	private String email;
+}
