+++ hiapp/src/main/java/com/hifive/bururung/domain/member/entity/MemberState.java
@@ -1,0 +1,6 @@
+package com.hifive.bururung.domain.member.entity;
+
+public enum MemberState {
+	ACTIVE,
+	INACTIVE
+}
