+++ hiapp/src/main/java/com/hifive/bururung/domain/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.hifive.bururung.domain.member.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.hifive.bururung.domain.member.entity.Member;
+
+public interface MemberRepository extends JpaRepository<Member, Long>{
+	Optional<Member> findByEmail(String email);
+}
