+++ hiapp/src/main/java/com/hifive/bururung/domain/member/service/IMemberService.java
@@ -1,0 +1,9 @@
+package com.hifive.bururung.domain.member.service;
+
+import com.hifive.bururung.domain.member.dto.SignupRequest;
+import com.hifive.bururung.domain.member.dto.TokenDTO;
+
+public interface IMemberService {
+	TokenDTO login(String email, String password);
+	Long signup(SignupRequest signupRequest);
+}
