--- hiapp/src/main/java/com/hifive/hiapp/HiappApplication.java
@@ -1,14 +1,0 @@
-package com.hifive.hiapp;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-@SpringBootApplication
-public class HiappApplication extends SpringBootServletInitializer {
-	
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
-		return builder.sources(HiappApplication.class);
-	}
-}
