--- hiapp/src/main/java/com/hifive/hiapp/domain/member/domain/MemberState.java
@@ -1,6 +1,0 @@
-package com.hifive.hiapp.domain.member.domain;
-
-public enum MemberState {
-	ACTIVE,
-	INACTIVE
-}
