--- hiapp/src/main/java/com/hifive/hiapp/domain/member/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.hifive.hiapp.domain.member.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.hifive.hiapp.domain.member.domain.Member;
-
-public interface MemberRepository extends JpaRepository<Member, Long>{
-	Optional<Member> findByEmail(String email);
-}
