--- hiapp/src/main/java/com/hifive/hiapp/domain/member/service/UserDetailsServiceImpl.java
@@ -1,35 +1,0 @@
-package com.hifive.hiapp.domain.member.service;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.hifive.hiapp.domain.member.domain.Member;
-import com.hifive.hiapp.domain.member.repository.MemberRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-	private final MemberRepository memberRepository;
-	
-	@Override
-	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new NoSuchElementException("사용자가 존재하지 않습니다."));
-        
-        List<GrantedAuthority> authorities = AuthorityUtils.createAuthorityList(member.getRoleName());
-
-        return new User(member.getEmail(), member.getPassword(), authorities);
-	}
-}
