--- hiapp/src/main/java/com/hifive/hiapp/global/common/CookieUtil.java
@@ -1,37 +1,0 @@
-package com.hifive.hiapp.global.common;
-
-import java.util.Optional;
-
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@Component
-public class CookieUtil {
-	
-	public void addCookie(HttpServletResponse response, String name, String value, int maxAge) {
-		Cookie cookie = new Cookie(name, value);
-		cookie.setPath("/");
-		cookie.setHttpOnly(true);
-		cookie.setMaxAge(maxAge);
-		cookie.setSecure(true);
-		
-		response.addCookie(cookie);
-	}
-	
-	public Optional<Cookie> getCookie(HttpServletRequest request, String name) {
-		Cookie[] cookies = request.getCookies();
-		
-		if(cookies != null) {
-			for(Cookie cookie : cookies) {
-				if(cookie.getName().equals(name)) {
-					return Optional.of(cookie);
-				}
-			}
-		}
-		
-		return Optional.empty();
-	}
-}
