+++ hiapp/src/main/resources/mapper/member/MemberMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper SYSTEM "file:///C:/dev/workspace/finalmeta/backend/hiapp/src/main/resources/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.hifive.hiapp.domain.member.repository.MemberMapper">
+<mapper namespace="com.hifive.bururung.domain.member.repository.MemberMapper">
     <!-- 전체 회원 조회 -->
-    <select id="findAll" resultType="com.hifive.hiapp.domain.member.dto.MemberDTO">
+    <select id="findAll" resultType="com.hifive.bururung.domain.member.dto.MemberDTO">
         SELECT 
             member_id AS memberId,
             email,
